More Navigation Progress
Removed titles from each section and added a div with fontawesome nav
arrows at the top and bottom of each section. Nav elements react when
div is in view now instead of h1.

Also, edited jQuery so border-bottom disappears when you scroll back up
to the html5 video at the top.

Also, made progress with “Back to Top” button. Still want to add
scrolling function to it instead of it just jumping to top.